recorder/requirements.txt
backcall==0.2.0
backports.zoneinfo==0.2.1
boto3==1.18.4
botocore==1.21.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
click-loglevel==0.4.0.post1
decorator==5.1.0
ffprobe-python==1.0.3
flake8==3.9.2
idna==3.2
ipython==7.27.0
jedi==0.18.0
jmespath==0.10.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.21.2
opencv-python==4.5.3.56
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
prompt-toolkit==3.0.20
psutil==5.8.0
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
pyudev==0.22.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sounddevice==0.4.2
toml==0.10.2
tomli==2.0.0
traitlets==5.1.0
typing-extensions==3.10.0.0
tzlocal==3.0
urllib3==1.26.6
v4l2ctl==0.1a5
wcwidth==0.2.5
